Utilisation du type uint32 et uint8 au lieu de int

+++ src/xor.c
@@ -5,8 +5,8 @@
 #include "../inc/utils.h"
 
 void xor(byte str_in[], byte str_out[], uint32 str_length, byte key[]) {
-    int k = 0;
-    for (int i = 0; i < str_length; ++i) {
+    uint8 k = 0;
+    for (uint32 i = 0; i < str_length; ++i) {
         if (key[k] == '\0') k = 0;
         str_out[i] = str_in[i] ^ key[k];
         ++k;
